refactor: change customer sales & estimates endpoint Restful resource - change `/api/sales?customerId=""` to `/api/customers/:id/sales` - update controller, service and repo layer
