+++ src/main/modules/customers/customers.types.ts
@@ -1,0 +1,11 @@
+export type PaginatedQuery = {
+  pageNo: number;
+  pageSize: number;
+};
+export type SalesByCustomerParams = PaginatedQuery & {
+  customerId: string;
+};
+
+export type EstimatesByCustomerParams = PaginatedQuery & {
+  customerId: string;
+};
