--- src/main/modules/estimates/estimates.types.ts
@@ -1,8 +1,0 @@
-export type PaginatedQuery = {
-  pageNo: number;
-  pageSize: number;
-};
-
-export type EstimatesByCustomerParams = PaginatedQuery & {
-  customerId: string;
-};
